[ #2083 ] Moved the internal tests to a Cabal test-suite (fourth part).
