+++ Agda.cabal
@@ -672,6 +672,7 @@ test-suite internal-tests
   main-is:         InternalTests.hs
 
   build-depends:  Agda == 2.5.2
+                , array >= 0.4.0.1 && < 0.6
                 , base >= 4.6.0.1 && < 4.10
                 , containers >= 0.5.0.0 && < 0.6
                 , haskell-src-exts >= 1.17 && < 1.18
