+++ src/full/Agda/Termination/CallMatrix.hs
@@ -20,7 +20,7 @@ import Data.Traversable (Traversable)
 
 import Agda.Termination.CutOff
 import Agda.Termination.Order as Order hiding (tests)
-import Agda.Termination.SparseMatrix as Matrix hiding (tests)
+import Agda.Termination.SparseMatrix as Matrix
 import Agda.Termination.Semiring (HasZero(..))
 
 import Agda.Utils.Favorites (Favorites)
@@ -99,7 +99,7 @@ type ArgumentIndex = Int
 
 
 newtype CallMatrix' a = CallMatrix { mat :: Matrix ArgumentIndex a }
-  deriving (Eq, Ord, Show, Functor, Foldable, Traversable, CoArbitrary, PartialOrd)
+  deriving (Eq, Ord, Show, Functor, Foldable, Traversable, PartialOrd)
 
 type CallMatrix = CallMatrix' Order
 
@@ -178,8 +178,7 @@ noAug m = CallMatrixAug m mempty
 -- | Sets of incomparable call matrices augmented with path information.
 --   Use overloaded 'null', 'empty', 'singleton', 'mappend'.
 newtype CMSet cinfo = CMSet { cmSet :: Favorites (CallMatrixAug cinfo) }
-  deriving ( Show, Arbitrary, CoArbitrary
-           , Semigroup, Monoid, Null, Singleton (CallMatrixAug cinfo) )
+  deriving ( Show, Semigroup, Monoid, Null, Singleton (CallMatrixAug cinfo) )
 
 -- | Call matrix set product is the Cartesian product.
 
@@ -216,84 +215,3 @@ instance Pretty cinfo => Pretty (CallMatrixAug cinfo) where
 instance Pretty cinfo => Pretty (CMSet cinfo) where
   pretty = vcat . punctuate newLine . map pretty . toList
     where newLine = text "\n"
-
-------------------------------------------------------------------------
--- * Generators and tests
-------------------------------------------------------------------------
-
--- ** CallMatrix
-
-instance Arbitrary CallMatrix where
-  arbitrary = callMatrix =<< arbitrary
-
--- | Generates a call matrix of the given size.
-
-callMatrix :: Size ArgumentIndex -> Gen CallMatrix
-callMatrix sz = CallMatrix <$> matrix sz
-
--- ** CallMatrixAug
-
-instance Arbitrary cinfo => Arbitrary (CallMatrixAug cinfo) where
-  arbitrary = CallMatrixAug <$> arbitrary <*> arbitrary
-
-instance CoArbitrary cinfo => CoArbitrary (CallMatrixAug cinfo) where
-  coarbitrary (CallMatrixAug m info) = coarbitrary m . coarbitrary info
-
-------------------------------------------------------------------------
--- * All tests
-------------------------------------------------------------------------
-
-tests :: IO Bool
-tests = runTests "Agda.Termination.CallMatrix"
-  [
-  ]
-  where ?cutoff = DontCutOff -- CutOff 2  -- don't cut off in tests!
-
-
--- RETIRED:  LONG OUTDATED call matrix invariant
-
--- -- | In a call matrix at most one element per row may be different
--- -- from 'unknown'.
-
--- callMatrixInvariant :: CallMatrix -> Bool
--- callMatrixInvariant (CallMatrix m) =
---   matrixInvariant m &&
---   all ((<= 1) . length . filter (/= unknown)) (toLists m)
-
--- prop_Arbitrary_CallMatrix = callMatrixInvariant
-
--- -- | Generates a call matrix of the given size.
-
--- callMatrix :: Size ArgumentIndex -> Gen CallMatrix
--- callMatrix sz = do
---   m <- matrixUsingRowGen sz rowGen
---   return $ CallMatrix { mat = m }
---   where
---   rowGen :: ArgumentIndex -> Gen [Order]
---   rowGen 0 = return []
---   rowGen n = do
---     x <- arbitrary
---     i <- choose (0, n - 1)
---     return $ genericReplicate i unknown ++ [x] ++
---              genericReplicate (n - 1 - i) unknown
-
--- prop_callMatrix sz =
---   forAll (callMatrix sz) $ \cm ->
---     callMatrixInvariant cm
---     &&
---     size (mat cm) == sz
-
--- prop_cmMul sz =
---   forAll natural $ \c2 ->
---   forAll (callMatrix sz) $ \cm1 ->
---   forAll (callMatrix $ Size { rows = cols sz, cols = c2 }) $ \cm2 ->
---     callMatrixInvariant (cm1 >*< cm2)
-
--- tests :: IO Bool
--- tests = runTests "Agda.Termination.CallMatrix"
---   [ quickCheck' callMatrixInvariant
---   , quickCheck' prop_Arbitrary_CallMatrix
---   , quickCheck' prop_callMatrix
---   , quickCheck' prop_cmMul
---   ]
---   where ?cutoff = DontCutOff -- CutOff 2  -- don't cut off in tests!
