+++ test/InternalTests.hs
@@ -12,11 +12,11 @@ import InternalTests.Interaction.Highlighting.Range           as InteRang     ( 
 import InternalTests.Interaction.Options                      as InteOpti     ( tests )
 import InternalTests.Syntax.Parser.Parser                     as SyntPars     ( tests )
 import Agda.Syntax.Position                                   as SyntPosi     ( tests )
-import Agda.Termination.CallGraph                             as TermCall     ( tests )
-import Agda.Termination.CallMatrix                            as TermCM       ( tests )
+import InternalTests.Termination.CallGraph                    as TermCall     ( tests )
+import InternalTests.Termination.CallMatrix                   as TermCM       ( tests )
 import Agda.Termination.Order                                 as TermOrd      ( tests )
 import Agda.Termination.Semiring                              as TermRing     ( tests )
-import Agda.Termination.SparseMatrix                          as TermSparse   ( tests )
+import InternalTests.Termination.SparseMatrix                 as TermSparse   ( tests )
 import Agda.Termination.Termination                           as TermTerm     ( tests )
 import InternalTests.TypeChecking.Free                        as Free         ( tests )
 import InternalTests.TypeChecking.Irrelevance                 as Irrel        ( tests )
